+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `seed-file` to `create` command [#226] 
 - Add `name` to `export` command 
 
 ### Fixed